requirements.txt
flask==2.1.2
pyyaml
